feat: refactor - rename yelp_repository to restaurant_repository
