+++ lib/features/restaurants_list/restaurants_list_screen.dart
@@ -7,14 +7,17 @@ import 'package:restaurantour/core/text_style.dart';
 import 'package:restaurantour/features/restaurants_list/pages/favorites/favorites_page.dart';
 import 'package:restaurantour/features/restaurants_list/pages/restaurants/restaurants_page.dart';
 import 'package:restaurantour/features/restaurants_list/restaurant_view_model.dart';
-import 'package:restaurantour/repositories/yelp_repository.dart';
+import 'package:restaurantour/repositories/restaurant_repository.dart';
 import 'package:restaurantour/services/favorites_service.dart';
 
 class RestaurantsListScreen extends StatefulWidget {
   const RestaurantsListScreen({super.key});
 
   static Widget create() => ChangeNotifierProvider(
-        create: (context) => RestaurantListViewModel(yelpRepo: inject<YelpRepository>(), favoritesService: inject<FavoritesService>()),
+        create: (context) => RestaurantListViewModel(
+          favoritesService: inject<FavoritesService>(),
+          restaurantRepository: inject<RestaurantRepository>(),
+        ),
         child: const RestaurantsListScreen(),
       );
